move openReportFromDeepLink to Report
